+++ src/Sudoku.Analytics/Analytics/Steps/DeadlyPatterns/Rectangles/UniqueRectangleExternalType4Step.cs
@@ -41,7 +41,7 @@ public sealed partial class UniqueRectangleExternalType4Step(
 	IIncompleteTrait
 {
 	/// <inheritdoc/>
-	public override decimal BaseDifficulty => base.BaseDifficulty + 2;
+	public override int BaseDifficulty => base.BaseDifficulty + 2;
 
 	/// <inheritdoc/>
 	public override FormatInterpolation[] FormatInterpolationParts
