+++ src/Sudoku.Analytics/Analytics/Steps/DeadlyPatterns/Reverse/ReverseBivalueUniversalGraveType2Step.cs
@@ -26,7 +26,7 @@ public sealed partial class ReverseBivalueUniversalGraveType2Step(
 	public override int Type => 2;
 
 	/// <inheritdoc/>
-	public override decimal BaseDifficulty => base.BaseDifficulty + 1;
+	public override int BaseDifficulty => base.BaseDifficulty + 1;
 
 	/// <inheritdoc/>
 	public override FormatInterpolation[] FormatInterpolationParts
