+++ src/Sudoku.Analytics/Analytics/Steps/Exocets/WeakExocetSlashStep.cs
@@ -25,7 +25,7 @@ public sealed partial class WeakExocetSlashStep(
 ) : ExocetStep(conclusions, views, options, digitsMask, in baseCells, in targetCells, [], in crosslineCells)
 {
 	/// <inheritdoc/>
-	public override decimal BaseDifficulty => base.BaseDifficulty + 8;
+	public override int BaseDifficulty => base.BaseDifficulty + 8;
 
 	/// <inheritdoc/>
 	public override Technique Code => Technique.WeakExocetSlash;
